# Ignore `total_order_seek` in DB::Get